Bug fixes, numpy type safety, Python 3.12 compat, aarch64 build
- atrip: fix numpy deprecation warnings (explicit dtype in np.arange,
  explicit int() casts for numpy scalars); handle str input in unlzw
- sawx: guard against None active_editor/focused_viewer in idle handler
  and frame title; use integer division in wx layout calculations;
  fix regex raw strings and escaped backslash in format strings
- omnivore: guard None focused_viewer in byte_editor; handle dict
  viewers in tile_manager_base_editor; bump version to 1.99.10
- Add check_deps script, default layout templates, aarch64 C extensions

+++ atrip/filesystems/atari_cas.py
@@ -84,7 +84,7 @@ class AtariCasDirent(Dirent):
                 if flag == 0xfe:
                     # end of file record
                     keep_reading = False
-                offsets[offset_index:offset_index + data_size] = np.arange(index, index + data_size)
+                offsets[offset_index:offset_index + data_size] = np.arange(index, index + data_size, dtype=offsets.dtype)
                 offset_index += data_size
             index += chunk.record_length
         self.temporary_offset_storage = np.copy(offsets[0:offset_index])
